Delete: Listing for `vid` extension
GitHub limits us to uploading files that are 100MB in size, but this
extension's bundle is 300MB now, with all of the cross-platform
libraries and stuff.

I'll need to do some work separately to figure out how to get this thing
down to a reasonable size.  (It might be that all of the JavaFX stuff is
just fluff.)